+++ Hybrid/vtkXYPlotActor.h
@@ -226,10 +226,10 @@ public:
   // plotted. Each input dataset (or data object) results in one curve. The
   // methods that follow have an index i that corresponds to the input dataset
   // or data object. 
-  void SetPlotColor(int i, float r, float g, float b);
-  void SetPlotColor(int i, const float color[3]) {
+  void SetPlotColor(int i, double r, double g, double b);
+  void SetPlotColor(int i, const double color[3]) {
     this->SetPlotColor(i, color[0], color[1], color[2]); };
-  float *GetPlotColor(int i);
+  double *GetPlotColor(int i);
   void SetPlotSymbol(int i,vtkPolyData *input);
   vtkPolyData *GetPlotSymbol(int i);
   void SetPlotLabel(int i, const char *label);
